toolkit/devtools/server/tests/mochitest/test_memory_allocations_04.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1068171 - Test controlling the memory actor's allocation sampling probability.
-->
<head>
  <meta charset="utf-8">
  <title>Memory monitoring actor test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="memory-helpers.js" type="application/javascript;version=1.8"></script>
<script>
window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  Task.spawn(function* () {
    var { memory, client } = yield startServerAndGetSelectedTabMemory();
    yield memory.attach();

    var allocs = [];
    function allocator() {
      for (var i = 0; i < 100; i++) {
        allocs.push({});
      }
    }

    var testProbability = Task.async(function* (p, expected) {
      info("probability = " + p);
      yield memory.startRecordingAllocations({
        probability: p
      });
      allocator();
      var response = yield memory.getAllocations();
      yield memory.stopRecordingAllocations();
      return response.allocations.length;
    });

    is((yield testProbability(0.0)), 0,
       "With probability = 0.0, we shouldn't get any allocations.");

    ok((yield testProbability(1.0)) >= 100,
       "With probability = 1.0, we should get all 100 allocations (plus "
       + "whatever allocations the actor and SpiderMonkey make).");

    // We don't test any other probabilities because the test would be
    // non-deterministic. We don't have a way to control the PRNG like we do in
    // jit-tests
    // (js/src/jit-test/tests/debug/Memory-allocationsSamplingProbability-*.js).

    yield memory.detach();
    destroyServerAndFinish(client);
  });
};
</script>
</pre>
</body>
</html>
